NLP Test First Working Version
This is the first real working version of the nlp test. It satisfies the
requirements for item 1 in the document. It performs word and then
sentence tokenization on an input stream of text. It prints the found
sentences to stdout in addition to and XML dump of the sentence
collection object model. The method that it uses is the first pass
annotation of the Punkt tokenization algorithm. If I have some
additional time I will add support for usupervised training of the Punkt
parameters and second phase annotation.